Merry Christmas python project for Beginners
Hello everyone , In this project I have told how to make a merry christmas animation card using the python programming language. The libraries used in this project are pygame, turtle, tie and random. So we had taken a gif image of christmas, a song for christmas and we had made a christmas tree and a stars using turtle in python. This is a python projects which is for beginners.